Add: README and comments on edgeidx
Separate contents of eisort.c from edgeidx.c
