+++ src/backend/access/edgeidx/Makefile
@@ -12,6 +12,6 @@ subdir = src/backend/access/edgeidx
 top_builddir = ../../../..
 include $(top_builddir)/src/Makefile.global
 
-OBJS = edgeidx.o eiinsert.o eisearch.o eipage.o
+OBJS = edgeidx.o eiinsert.o eisearch.o eipage.o eisort.o
 
 include $(top_srcdir)/src/backend/common.mk
